feat: move VersionedStorageScoringMetrics to iota_types
